feat: Add auto-commit alias to bashrc in install script
The modification allows for automatic git commit aliasing to be persisted across shells by adding it inside .bashrc file. Now, the 'auto-commit' command will be available once the bashrc file is sourced, streamlining the commit process.🐚🚀

In the original script, the git alias 'auto-commit' was configured but non-persistent as it would be lost once the shell is closed.

In the updated script, the 'auto-commit' alias is written into the user's .bashrc file, making it permanent. For immediate availability within the current session, the .bashrc file is sourced right after the alias is written into it.🛠️🌐

This will ease out the task for users who repeatedly use the 'auto-commit' command. Now, with this change, they can use this command directly across all bash shells, enhancing convenience and productivity.👨‍💻✨